requirements.txt
GitPython
requests
PyYaml
kubernetes
docker